+++ stats/src/main/java/com/facebook/airlift/stats/cardinality/SparseHll.java
@@ -40,7 +40,7 @@ import static java.util.Comparator.comparingInt;
 final class SparseHll
         implements HllInstance
 {
-    private static final int SPARSE_INSTANCE_SIZE = ClassLayout.parseClass(SparseHll.class).instanceSize();
+    private static final long SPARSE_INSTANCE_SIZE = ClassLayout.parseClass(SparseHll.class).instanceSize();
 
     // 6 bits to encode the number of zeros after the truncated hash
     // and be able to fit the encoded value in an integer
@@ -192,7 +192,7 @@ final class SparseHll
     }
 
     @Override
-    public int estimatedInMemorySize()
+    public long estimatedInMemorySize()
     {
         return SPARSE_INSTANCE_SIZE + toIntExact(sizeOf(entries));
     }
